Support create and enter room networking

+++ texas-holdem-common/src/lib.rs
@@ -8,7 +8,7 @@ pub const PROTOCOL_ID: u64 = 0;
 #[derive(Debug, Serialize, Deserialize)]
 pub struct RoomDTO {
     // 房间id
-    pub room_id: u32,
+    pub room_id: u64,
     // 房间名称
     pub room_name: String,
     // 房主名称
